refactor: address-form, main component 삭제 및 page에서 api 호출
